pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = [
  "/test_images",
  "/test_images_ex",
  "/hybrid_setup"
]

[project]
name = "vsdeoldify"
version = "4.5.0"
description = "HAVC functions for VapourSynth"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name = "Dan64", email = "whitedan64@gmail.com"}]
keywords = ["Deoldify", "VapourSynth"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Multimedia :: Video"
]
dependencies = [   
  "numpy>=1.26.2",
  "vsddcolor>=1.0.0",
  "fastai==1.0.60",
  "torch>=2.1.1",
  "pillow>=9.0.0",
  "VapourSynth>=62",
  "vstools>=2.1.0"
]

[project.urls]
"Homepage" = "https://github.com/dan64/vs-deoldify"
"Bug Tracker" = "https://github.com/dan64/vs-deoldify/issues"
